assistantgemini/requirements.txt
google-generativeai>=0.8.0
aiocache
chromadb
json5
msgpack
numpy
pandas
pydantic
pytz
sentry_sdk
tenacity
ujson
httpx 